move EmptyBitmap to tests and remove it from the API, fix bitmap64 countZeros
