+++ hdt-java-core/src/main/java/org/rdfhdt/hdt/compact/bitmap/Bitmap64.java
@@ -229,7 +229,7 @@ public class Bitmap64 implements ModifiableBitmap{
 
 	@Override
 	public long countZeros() {
-		return getRealSizeBytes() - countOnes();
+		return words.length * 64L - countOnes();
 	}
 
 	@Override
